requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-socketio==5.11.0
jinja2==3.1.3
pydantic==2.6.0
python-multipart==0.0.9